requirements.txt
streamlit>=1.28
st-gsheets-connection
gspread
google-auth
pyyaml
